media: rockchip: isp1: stop isp when too many errors are reported
when use hdmi-in chip with isp, it needs to pull out/plug in hdmi cable
when stream is on.
many error logs are reported in interrupt function when pull out/plug in
hdmi cable, then the system will be dead.
so we stop isp when too many error logs are reported.

Change-Id: Ic19eddfc952954a11a176b91c16ce4c9b6372418
Signed-off-by: Hu Kejun <william.hu@rock-chips.com>